code cleanup and tests fixed
