+++ src/test-gui/java/de/qabel/desktop/ui/contact/ContactGuiTest.java
@@ -1,15 +1,23 @@
 package de.qabel.desktop.ui.contact;
 
 import com.airhacks.afterburner.views.FXMLView;
+import com.sun.deploy.uitoolkit.Window;
 import de.qabel.core.config.Contact;
 import de.qabel.core.config.Identity;
 import de.qabel.core.repository.exception.PersistenceException;
 import de.qabel.desktop.ui.AbstractGuiTest;
+import javafx.scene.Scene;
+import javafx.scene.layout.Pane;
+import javafx.scene.paint.Color;
+import javafx.stage.Stage;
 import org.junit.Test;
 
 import static de.qabel.desktop.AsyncUtils.assertAsync;
+import static de.qabel.desktop.AsyncUtils.runLaterAndWait;
+import static de.qabel.desktop.AsyncUtils.waitUntil;
 import static junit.framework.TestCase.assertEquals;
 import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertTrue;
 
 public class ContactGuiTest extends AbstractGuiTest<ContactController> {
 
@@ -42,6 +50,16 @@ public class ContactGuiTest extends AbstractGuiTest<ContactController> {
         Identity i3 = identityBuilderFactory.factory().withAlias("I" + name).build();
         Contact c3 = new Contact("C" + name, i3.getDropUrls(), i3.getEcPublicKey());
         contactRepository.save(c3, identity);
+    }
+
+    @Test
+    public void testOpenMenu() {
+        ContactPage page = new ContactPage(baseFXRobot, robot, controller);
+        controller.layoutWindow = controller.contactList;
+        page.openMenu();
+        waitUntil(() -> controller.contactMenuController != null);
+        waitUntil(() -> controller.contactMenuController.popOver != null);
+        assertAsync(() -> controller.contactMenuController.popOver.isShowing());
     }
 
     @Test
